fix: set storage column type to BIGINT
set default storage limit if StorageLimit is empty/0

+++ server/migrations/96_storage_table_families.up.sql
@@ -1,2 +1,2 @@
 ALTER TABLE families
-ADD COLUMN storage INTEGER;
+ADD COLUMN storage BIGINT;
